backend/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
flask-cors
google-adk
google-adk
firebase-admin
requests>=2.25.0
pymongo>=4.0.0
dnspython>=2.0.0

# Production server dependencies
waitress>=3.0.0          # Windows-compatible WSGI server
gunicorn>=21.0.0         # Linux production server (use on deployment)
gevent>=24.0.0           # Async workers for gunicorn

# Performance & caching
cachetools>=5.3.0        # TTL caching for API responses
flask-limiter>=3.5.0     # Rate limiting
